refactored qnd resolved conflicts
