+++ app/src/main/java/com/amalitech/bookmeetingroom/BookMeetingApplication.kt
@@ -13,12 +13,8 @@ import com.amalitech.onboarding.di.onboardingDomainModule
 import com.amalitech.onboarding.di.onboardingPresentationModule
 import com.amalitech.onboarding_data.di.onboardingDataModule
 import com.amalitech.rooms.di.dataRoomModule
-import com.amalitech.rooms.di.domainRoomModule
+import com.amalitech.rooms.di.domainRoomsModule
 import com.amalitech.rooms.di.roomPresentationModule
-import com.amalitech.room.di.domainRoomModule
-import com.example.room.di.uiRoomModule
-import com.amalitech.room.di.domainRoomModule
-import com.amalitech.room.di.uiRoomModule
 import org.koin.android.ext.koin.androidContext
 import org.koin.core.context.GlobalContext.startKoin
 
@@ -36,16 +32,13 @@ class BookMeetingApplication : Application() {
                 appModule,
                 roomPresentationModule,
                 dataRoomModule,
-                domainRoomModule,
                 adminPresentationModule,
                 domainHomeModule,
                 uiHomeModule,
                 domainBookingModule,
-                presentationBookingModule
+                presentationBookingModule,
                 uiHomeModule,
-                domainRoomModule,
-                domainRoomModule,
-                uiRoomModule
+                domainRoomsModule,
             )
         }
     }
