+++ app/src/main/java/com/amalitech/bookmeetingroom/MainActivity.kt
@@ -6,15 +6,16 @@ import androidx.activity.compose.setContent
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.tooling.preview.Preview
+import androidx.navigation.compose.rememberNavController
+import com.amalitech.bookmeetingroom.navigation.NavigationGraph
 import com.amalitech.core_ui.theme.BookMeetingRoomTheme
-import com.amalitech.rooms.RoomListScreen
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContent {
             BookMeetingRoomTheme {
-               RoomListScreen(onNavigateToAddRoom = {})
+               NavigationGraph(navController = rememberNavController())
             }
         }
     }
