+++ domain/rooms/src/main/java/com/amalitech/rooms/book_room/model/Room.kt
@@ -1,4 +1,4 @@
-package com.amalitech.room.book_room.model
+package com.amalitech.rooms.book_room.model
 
 import com.amalitech.core.domain.model.Booking
