+++ domain/rooms/src/main/java/com/amalitech/rooms/di/domainRoomModule.kt
@@ -1,12 +1,15 @@
 package com.amalitech.rooms.di
 
+import com.amalitech.rooms.book_room.use_case.BookRoomUseCase
+import com.amalitech.rooms.book_room.use_case.BookRoomUseCasesWrapper
+import com.amalitech.rooms.book_room.use_case.GetRoomUseCase
 import com.amalitech.rooms.usecase.DeleteRoomUseCase
 import com.amalitech.rooms.usecase.FetchRoomsUseCase
 import com.amalitech.rooms.usecase.RoomUseCaseWrapper
 import org.koin.dsl.module
 
 
-val domainRoomModule = module {
+val domainRoomsModule = module {
     single {
         FetchRoomsUseCase(get())
     }
@@ -17,5 +20,17 @@ val domainRoomModule = module {
 
     single {
         RoomUseCaseWrapper(get(), get())
+    }
+
+    single {
+        GetRoomUseCase()
+    }
+
+    single {
+        BookRoomUseCase()
+    }
+
+    single {
+        BookRoomUseCasesWrapper(get(), get(), get())
     }
 }
