+++ settings.gradle
@@ -33,5 +33,3 @@ include ':ui:home'
 include 'domain:home'
 include ':ui:booking'
 include ':domain:booking'
-include ':ui:room'
-include ':domain:room'
