+++ ui/rooms/build.gradle.kts
@@ -3,8 +3,9 @@ apply {
 }
 dependencies {
     "implementation"(project(Modules.coreUI))
-    "implementation"(project(Modules.domainRoom))
+    "implementation"(project(Modules.domainRooms))
     "implementation"(project(Modules.core))
+    "implementation"(DatePicker.vanpra)
     "implementation"(Coil.coilCompose)
     "implementation"(Kotlin.coreKtx)
