+++ ui/rooms/src/main/java/com/amalitech/rooms/book_room/TimeUi.kt
@@ -1,4 +1,4 @@
-package com.amalitech.room.book_room
+package com.amalitech.rooms.book_room
 
 import java.time.LocalTime
