+++ ui/rooms/src/main/java/com/amalitech/rooms/book_room/components/FeatureItem.kt
@@ -1,9 +1,8 @@
-package com.amalitech.room.book_room.components
+package com.amalitech.rooms.book_room.components
 
+import androidx.compose.foundation.background
 import androidx.compose.foundation.border
-import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.MaterialTheme
@@ -12,38 +11,37 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
-import androidx.compose.ui.graphics.Shape
+import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.unit.dp
 import com.amalitech.core_ui.theme.LocalSpacing
 
 @Composable
-fun SelectDateBox(
+fun FeatureItem(
+    feature: String,
     modifier: Modifier = Modifier,
-    onClick: () -> Unit,
-    text: String,
+    borderColor: Color = MaterialTheme.colorScheme.primary,
+    backgroundColor: Color = MaterialTheme.colorScheme.background,
+    contentColor: Color = MaterialTheme.colorScheme.onBackground,
     textStyle: TextStyle = MaterialTheme.typography.bodyMedium,
-    shape: Shape = RoundedCornerShape(LocalSpacing.current.spaceExtraSmall)
 ) {
     val spacing = LocalSpacing.current
     Box(
         modifier = modifier
-            .fillMaxSize()
-            .clip(RoundedCornerShape(spacing.spaceExtraSmall))
+            .clip(RoundedCornerShape(spacing.spaceMedium))
             .border(
-                1.dp,
-                MaterialTheme.colorScheme.onBackground,
-                shape = shape
+                color = borderColor,
+                width = 1.dp,
+                shape = RoundedCornerShape(spacing.spaceMedium)
             )
-            .clickable {
-                onClick()
-            },
-        contentAlignment = Alignment.Center
+            .background(backgroundColor)
+            .padding(spacing.spaceSmall),
+        contentAlignment = Alignment.CenterStart
     ) {
         Text(
-            text = text,
+            text = feature,
             style = textStyle,
-            modifier = Modifier.padding(spacing.spaceExtraSmall)
+            color = contentColor
         )
     }
 }
