+++ ui/rooms/src/main/java/com/amalitech/rooms/book_room/components/SelectDateBox.kt
@@ -1,8 +1,9 @@
-package com.amalitech.room.book_room.components
+package com.amalitech.rooms.book_room.components
 
-import androidx.compose.foundation.background
 import androidx.compose.foundation.border
+import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.MaterialTheme
@@ -11,37 +12,38 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
-import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.Shape
 import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.unit.dp
 import com.amalitech.core_ui.theme.LocalSpacing
 
 @Composable
-fun FeatureItem(
-    feature: String,
+fun SelectDateBox(
     modifier: Modifier = Modifier,
-    borderColor: Color = MaterialTheme.colorScheme.primary,
-    backgroundColor: Color = MaterialTheme.colorScheme.background,
-    contentColor: Color = MaterialTheme.colorScheme.onBackground,
+    onClick: () -> Unit,
+    text: String,
     textStyle: TextStyle = MaterialTheme.typography.bodyMedium,
+    shape: Shape = RoundedCornerShape(LocalSpacing.current.spaceExtraSmall)
 ) {
     val spacing = LocalSpacing.current
     Box(
         modifier = modifier
-            .clip(RoundedCornerShape(spacing.spaceMedium))
+            .fillMaxSize()
+            .clip(RoundedCornerShape(spacing.spaceExtraSmall))
             .border(
-                color = borderColor,
-                width = 1.dp,
-                shape = RoundedCornerShape(spacing.spaceMedium)
+                1.dp,
+                MaterialTheme.colorScheme.onBackground,
+                shape = shape
             )
-            .background(backgroundColor)
-            .padding(spacing.spaceSmall),
-        contentAlignment = Alignment.CenterStart
+            .clickable {
+                onClick()
+            },
+        contentAlignment = Alignment.Center
     ) {
         Text(
-            text = feature,
+            text = text,
             style = textStyle,
-            color = contentColor
+            modifier = Modifier.padding(spacing.spaceExtraSmall)
         )
     }
 }
