+++ ui/rooms/src/main/java/com/amalitech/rooms/book_room/util/NavArguments.kt
@@ -1,4 +1,4 @@
-package com.amalitech.room.book_room.util
+package com.amalitech.rooms.book_room.util
 
 object NavArguments {
     const val roomId = "roomId"
