+++ ui/rooms/src/main/res/values/strings.xml
@@ -1,7 +1,28 @@
 <resources>
     <string name="app_name">Rooms</string>
     <string name="up_to_people">up to %d people</string>
-    <string name="delete">Delete</string>
-    <string name="cancel">Cancel</string>
     <string name="question_delete_room">Are you sure you want to delete %s Room</string>
+    <string name="error_no_date_selected">No date selected, please select one before proceeding</string>
+    <string name="ok">OK</string>
+    <string name="cancel">Cancel</string>
+    <string name="select_time">Select time</string>
+    <string name="start">Start</string>
+    <string name="features">Features</string>
+    <string name="attendees">Attendees</string>
+    <string name="add_attendee">Add attendee</string>
+    <string name="delete">Delete</string>
+    <string name="error_no_attendee_added">No attendee added, please provide at least one attendee</string>
+    <string name="error_provide_an_end_time">Please provide an end time for the meeting</string>
+    <string name="error_provide_start_time">Please enter the start time</string>
+    <string name="error_provide_date_for_the_meeting">Please provide a date for the meeting</string>
+    <string name="note">Note</string>
+    <string name="book">Book</string>
+    <string name="error_select_start_time">Please select start time and date first</string>
+    <string name="end">End</string>
+    <string name="error_select_date">Please select the date first</string>
+    <string name="select_date">Select date</string>
+    <string name="pick_date">Pick a date</string>
+    <string name="x_more">&amp; %d more</string>
+    <string name="selected_date">Selected date %s</string>
+    <string name="select_date_and_time">Select date and time</string>
 </resources>
